refactor(CO-3211): update participants type imports to use specific paths for better clarity
