+++ src/__test__/generators/getMsgResponse.ts
@@ -8,7 +8,9 @@ import { faker } from '@faker-js/faker';
 import { SoapResponse } from '@zextras/carbonio-shell-ui';
 import { FOLDERS, ParticipantRole } from '@zextras/carbonio-ui-commons';
 
-import type { GetMsgResponse, Participant, SoapMailParticipant } from 'types';
+import { Participant } from 'types/participant';
+import { GetMsgResponse } from 'types/soap/get-msg';
+import { SoapMailParticipant } from 'types/soap/soap-mail-participant';
 
 /**
  *
