+++ src/helpers/identities.ts
@@ -12,7 +12,8 @@ import { filter, findIndex, flatten, map, remove } from 'lodash';
 import { NO_ACCOUNT_NAME } from 'constants/index';
 import { getFolderIdParts, getMessageOwnerAccountName } from 'helpers/folders';
 import { getAvailableAddresses } from 'helpers/get-available-addresses';
-import type { MailMessage, Participant } from 'types/index.d';
+import { MailMessage } from 'types/messages';
+import { Participant } from 'types/participant';
 
 /**
  * The name of the primary identity
