+++ src/integrations/shared-functions.ts
@@ -6,7 +6,7 @@
 
 import { EditViewActions } from 'constants/index';
 import { EditorPrefillData } from 'types/editor';
-import type { Participant } from 'types/index.d';
+import { Participant } from 'types/participant';
 import { createEditBoard } from 'views/app/detail-panel/edit/edit-view-board';
 
 export const mailToSharedFunction: (recipients: Array<Participant>, subject?: string) => void = (
