+++ src/types/index.d.ts
@@ -9,5 +9,4 @@ export * from './api';
 export * from './attachments';
 export * from './conversations';
 export * from './messages';
-export * from './participant';
 export * from './soap';
