+++ src/ui-actions/participant-displayer-actions.ts
@@ -7,7 +7,7 @@ import { CreateSnackbarFn } from '@zextras/carbonio-design-system';
 import { t } from '@zextras/carbonio-shell-ui';
 
 import { mailToSharedFunction } from 'integrations/shared-functions';
-import { Participant } from 'types/index.d';
+import { Participant } from 'types/participant';
 
 export const copyEmailToClipboard = (email: string, createSnackbar: CreateSnackbarFn): void => {
 	navigator.clipboard.writeText(email).then(() => {
